make Temporary method private
